feat(web): Added calendar reservation room view connected to backend
