+++ apps/web/app/(host)/hosting/reservations/calendar/properties/rooms/page.tsx
@@ -2,7 +2,7 @@ import { HOST } from "@/common/constants"
 import { APP_NAME } from "@repo/constants"
 import { Metadata } from "next"
 import AuthGuard from "@/common/components/AuthGuard"
-import CalendarProperty from "@/module/Hosting/Reservations/Calendar/Property"
+import CalendarProperty from "@/module/Hosting/Reservations/Calendar/Property/RoomReservationCalendar.tsx"
 
 export const metadata: Metadata = {
   title: `${HOST} - ${APP_NAME}`,
