+++ apps/web/module/Hosting/Reservations/Calendar/Property/RoomReservationCalendar.tsx.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { Typography } from "@/common/components/ui/Typography"
-import CalendarTable from "./CalendarTable"
+import CalendarTable from "./RoomCalendarTable"
 import { Table } from "lucide-react"
 import { Button } from "@/common/components/ui/Button"
 import { useRouter } from "next/navigation"
