+++ apps/web/module/Hosting/Reservations/helpers/reservationCalendarTabs.tsx
@@ -9,7 +9,7 @@ const reservationCalendarTabs = [
   {
     name: "Properties",
     icon: <LucideBuilding2 className="w-5" />,
-    link: "/hosting/reservations/calendar/properties",
+    link: "/hosting/reservations/calendar/properties/rooms",
   },
   {
     name: "Rentals",
